Lots of changes were made
1.  The basic architecture was changed from using Collections to basic arrays
2.  Progress bar was updated and hopefully fixed
3.  Code and form objects were condensed

For the architecture, the code uses a 2D array which is bound by the date selection and the amount of workers.  The rows are each date while the columns are workers.  By using a 2D array it made it simple to create a matrix of work times that can be manipulated easily and also saves a little memory.  All Dictionary Collections were replaced with arrays and there are helper methods for finding indexes.  All of the indexes coorespond to the date indexes.  Only the workers array index is unique.  It is a 2D array of WorkingTimes which is a class dedicated to holding all event names, event start and end times, as well as overlapping flags.  Data is found by finding the index of a date within selection and then finding the index of a specific worker.  The data includes every event and work times of a worker through arrays established in the WorkingTime class.

For example, if I was working on a date where the index was 2 and my personal work index was 43 then:
workingTimes[2, 43] will access the WorkingTime object that is in relation to those two parameters.  From there, data on every event I was signed up for as well as start and end times can be accessed.  All the arrays within the class use the same indexing method so event details and working times will both be on the same index to connnect the two.

4.  Final update was mentioned in 3 a little.  I took out some buttons like the update button for the chart and replaced the update process with simply clicking the name.

Everything should be working.  I checked it with the cases I had and it worked.  Also no need for the Papera edit and Z staff check because my new code accounts for those problems.